+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Implement snd_async_handler_get_callback_private
 * Allow users to resize analog columns in input editor
 * Add more options to lua gui.text
+* Add lua memory.readcstring()
 
 ### Changed